feat(archive): implement archive extraction functionality for tar.gz and tar.xz formats
